requirements.txt
langchain
langchain-core
langchain-openai
langgraph
chromadb
pydantic
python-dotenv
